ci-scripts/infra-setup/roles/rrcockpit/files/telegraf_py3/requirements.txt
bs4
click
dlrnapi-client
requests
rich
PyYAML
